\name{authenticate}
\alias{authenticate}
\title{authenticate}
\description{Authenticate with GSCF}
\usage{authenticate(user, pass, shared.key)}
\arguments{
  \item{user}{The username}
  \item{pass}{The password}
  \item{shared.key}{The shared key (can be found at your profile page in the GSCF web interface)}
}



\author{Thomas Kelder <thomas.kelder@tno.nl>}





