\name{getAssaysForStudy}
\alias{getAssaysForStudy}
\title{getAssaysForStudy}
\description{Get the available assays for a given study}
\usage{getAssaysForStudy(studyToken)}
\arguments{
  \item{studyToken}{The token of the study to get the assays for}
}

\value{A named list of the assays}

\author{Thomas Kelder, Ferry Jagers}





