\name{getEventGroupsForStudy}
\alias{getEventGroupsForStudy}
\title{getEventGroupsForStudy}
\description{Get event groups for a given study}
\usage{getEventGroupsForStudy(studyToken)}
\arguments{
  \item{studyToken}{The token of the study to get the event groups for}
}

\value{A named list with the event groups}

\author{Thomas Kelder <thomas.kelder@tno.nl>}





