\name{getEventsForStudy}
\alias{getEventsForStudy}
\title{getEventsForStudy}
\description{Get events for a given study}
\usage{getEventsForStudy(studyToken)}
\arguments{
  \item{studyToken}{The token of the study to get the events for}
}

\value{A named list with the events}

\author{Thomas Kelder <thomas.kelder@tno.nl>}





