\name{getFeaturesForAssay}
\alias{getFeaturesForAssay}
\title{getFeaturesForAssay}
\description{Get features for a given assay}
\usage{getFeaturesForAssay(assayToken)}
\arguments{
  \item{assayToken}{The token of the assay to get the features for}
}

\value{A named list of the available features}

\author{Thomas Kelder, Ferry Jagers}





