\name{getMeasurementDataForAssay}
\alias{getMeasurementDataForAssay}
\title{getMeasurementDataForAssay}
\description{Get data for a given assay}
\usage{getMeasurementDataForAssay(assayToken)}
\arguments{
  \item{assayToken}{The token of the assay to get the data for}
}

\value{Named list with available measurement data}

\author{Thomas Kelder, Ferry Jagers}





