\name{getSampleAndTreatmentGroupsForStudy}
\alias{getSampleAndTreatmentGroupsForStudy}
\title{getSampleAndTreatmentGroupsForStudy}
\description{Get sample & treatment groups for a given study}
\usage{getEventGroupsForStudy(studyToken)}
\arguments{
  \item{studyToken}{The token of the study to get the sample & treatment groups for}
}

\value{A named list with the sample & treatment groups}

\author{Thomas Kelder, Ferry Jagers}





