\name{getSamplesForAssay}
\alias{getSamplesForAssay}
\title{getSamplesForAssay}
\description{Get samples for a given assay}
\usage{getSamplesForAssay(assayToken)}
\arguments{
  \item{assayToken}{The token of the assay to get the samples for}
}

\value{A named list of the available samples}

\author{Thomas Kelder, Ferry Jagers}





