\name{getSamplesForStudy}
\alias{getSamplesForStudy}
\title{getSamplesForStudy}
\description{Get samples for a given study}
\usage{getSamplesForStudy(studyToken)}
\arguments{
  \item{studyToken}{The token of the study to get the samples for}
}

\value{A named list with the samples}

\author{Thomas Kelder <thomas.kelder@tno.nl>}





