\name{getSampleTypesForStudy}
\alias{getSampleTypesForStudy}
\title{getSampleTypesForStudy}
\description{Get sample types for a given study}
\usage{getSampleTypesForStudy(studyToken)}
\arguments{
  \item{studyToken}{The token of the study to get the sample types for}
}

\value{A named list with the sample types}

\author{Thomas Kelder, Ferry Jagers}





