\name{getSamplingEventsForStudy}
\alias{getSamplingEventsForStudy}
\title{getSamplingEventsForStudy}
\description{Get sampling events for a given study}
\usage{getSamplingEventsForStudy(studyToken)}
\arguments{
  \item{studyToken}{The token of the study to get the sampling events for}
}

\value{A named list with the sampling events}

\author{Thomas Kelder <thomas.kelder@tno.nl>}





