\name{getSubjectsForAssay}
\alias{getSubjectsForAssay}
\title{getSubjectsForAssay}
\description{Get subjects for a given assay}
\usage{getSubjectsForAssay(assayToken)}
\arguments{
  \item{assayToken}{The token of the assay to get the subjects for}
}

\value{A named list of the available subjects}

\author{Thomas Kelder, Ferry Jagers}





