\name{getSubjectsForStudy}
\alias{getSubjectsForStudy}
\title{getSubjectsForStudy}
\description{Get subjects for a given study}
\usage{getSubjectsForStudy(studyToken)}
\arguments{
  \item{studyToken}{The token of the study to get the subjects for}
}

\value{A named list with the subjects}

\author{Thomas Kelder, Ferry Jagers}





