\name{getTreatmentTypesForStudy}
\alias{getTreatmentTypesForStudy}
\title{getTreatmentTypesForStudy}
\description{Get treatment types for a given study}
\usage{getTreatmentTypesForStudy(studyToken)}
\arguments{
  \item{studyToken}{The token of the study to get the treatment types for}
}

\value{A named list with the treatment types}

\author{Thomas Kelder, Ferry Jagers}




