\name{GSCFClient-package}
\alias{GSCFClient-package}
\alias{GSCFClient}
\docType{package}
\title{Client for GSCF (http://phenotypefoundation.org/)}
\description{Client for the Generic Study Capture Framework (GSCF, see http://phenotypefoundation.org/). Allows querying study description and underlying datasets.}
\details{
\tabular{ll}{Package: \tab GSCFClient\cr
Type: \tab Package\cr
Title: \tab Client for GSCF (http://phenotypefoundation.org/)\cr
Version: \tab 1.2\cr
Date: \tab 2012-07-18\cr
Author: \tab Thomas Kelder <thomas.kelder@tno.nl>\cr
Maintainer: \tab Thomas Kelder <thomas.kelder@tno.nl>\cr
License: \tab Apache License, Version 2.0\cr
Depends: \tab RJSONIO, RCurl, digest\cr}
}
\author{Thomas Kelder <thomas.kelder@tno.nl>}

\keyword{ package }


