\name{PhenotypeDatabase-RClient-package}
\alias{PhenotypeDatabase-RClient-package}
\alias{PhenotypeDatabase-RClient}
\docType{package}
\title{Client for the Phenotype Database (http://phenotypefoundation.org/)}
\description{Client for the Phenotype Database (phenotypefoundation.org). Allows querying study description and underlying datasets.}
\details{
\tabular{ll}{Package: \tab PhenotypeDatabase-RClient\cr
Type: \tab Package\cr
Title: \tab R Client for Phenotype Database (http://phenotypefoundation.org/)\cr
Version: \tab 2.0.0\cr
Date: \tab 2016-05-20\cr
Author: \tab Thomas Kelder, Ferry Jagers\cr
Maintainer: \tab Ferry Jagers <ferry.jagers@tno.nl>\cr
License: \tab Apache License, Version 2.0\cr
Depends: \tab RJSONIO, RCurl, digest\cr}
}
\author{Thomas Kelder, Ferry Jagers}

\keyword{ package }


