\name{authenticate}
\alias{authenticate}
\title{authenticate}
\description{Authenticate with PhenotypeDabase}
\usage{authenticate(user, pass, shared.key)}
\arguments{
  \item{user}{The username}
  \item{pass}{The password}
  \item{shared.key}{The shared key (can be found at your profile page in the Phenotype Database web interface)}
}

\author{Thomas Kelder, Ferry Jagers}



